+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](https://semver.org/) since version
   - `KeyboardEvent.pressed`, `KeyboardEvent.which`, `KeyboardEvent.window_id` attributes.
   - `WindowEvent.data` and `WindowEvent.window_id` attributes and added missing SDL3 window events.
   - `which` and `window_id` attributes for mouse events.
+  - Events now have `Event.timestamp` and `Event.timestamp_ns` which use SDL's timer at `tcod.event.time` and `tcod.event.time_ns`.
 
 ### Changed